fix: move resolveWidgetColor to leaf to avoid new cyclic deps
The dark-mode rewrite imported resolveWidgetColor from widgets/WidgetUtils into
JSONForm component files; WidgetUtils imports a JSONForm type, so those imports
closed two new import cycles (dpdm 2134 -> 2136). Move resolveWidgetColor into
the dependency-free constants/darkModeColors leaf (re-exported from WidgetUtils
for compat) and import it from the leaf in the widget components. dpdm circular
count back to the 2134 base; check-types and eslint clean.
