+++ app/client/src/widgets/ContainerWidget/component/index.tsx
@@ -11,9 +11,9 @@ import fastdom from "fastdom";
 import { generateClassName, getCanvasClassName } from "utils/generators";
 import type { WidgetStyleContainerProps } from "components/designSystems/appsmith/WidgetStyleContainer";
 import WidgetStyleContainer from "components/designSystems/appsmith/WidgetStyleContainer";
-import { resolveWidgetColor, scrollCSS } from "widgets/WidgetUtils";
+import { scrollCSS } from "widgets/WidgetUtils";
 import { useColorMode } from "widgets/ColorModeContext";
-import { DARK_MODE_COLORS } from "constants/darkModeColors";
+import { DARK_MODE_COLORS, resolveWidgetColor } from "constants/darkModeColors";
 import { useSelector } from "react-redux";
 import { LayoutSystemTypes } from "layoutSystems/types";
 import { MAIN_CONTAINER_WIDGET_ID } from "constants/WidgetConstants";
